+++ bot/cogs/common.py
@@ -3,9 +3,10 @@ from logging import getLogger
 
 import modules.config as cfg
 import modules.lobby as lobby
+from cogs.admin import get_check_player
 
 from match.classes.match import Match
-from display import AllStrings as disp
+from display import AllStrings as disp, ContextWrapper
 from match import MatchStatus
 
 log = getLogger("pog_bot")
@@ -34,6 +35,32 @@ class MatchesCog(commands.Cog, name='common'):
             await match.command.info(ctx)
             return
         await disp.WRONG_CHANNEL_2.send(ctx, ctx.command.name, f"<#{ctx.channel.id}>")
+
+    @commands.command(aliases=['rm'])
+    @commands.guild_only()
+    async def remove(self, ctx):
+        if ctx.channel.id == cfg.channels["lobby"]:
+            player = await get_check_player(ctx)
+            if not player:
+                return
+            if player.is_lobbied:
+                lobby.remove_from_lobby(player)
+                await disp.RM_LOBBY.send(ContextWrapper.channel(cfg.channels["lobby"]), player.mention,
+                                         names_in_lobby=lobby.get_all_names_in_lobby())
+                return
+            await disp.RM_NOT_LOBBIED.send(ctx)
+            return
+        # if ctx.channel.id == cfg.channels["register"]:
+        #     player = await get_check_player(ctx)
+        #     if not player:
+        #         return
+        #     #TODO: remove ig names
+        if ctx.channel.id in cfg.channels["matches"]:
+            match = Match.get(ctx.channel.id)
+            await match.command.bench(ctx)
+            return
+        else:
+            await disp.WRONG_CHANNEL_2.send(ctx, ctx.command.name, f"<#{ctx.channel.id}>")
 
 
 def setup(client):