+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Changed the way DM messages are handled
 - updated = timeout command
 - It is now possible to use =swap and =sub between two rounds
+- Added =remove in match channels: remove two players from match
 
 # v3.2:
 - Added locking mechanic on reaction handler to avoid errors due to players spamming the reactions